requirements.txt
blinker==1.9.0
click==8.3.1
croniter==6.0.0
Flask==3.1.2
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
playwright==1.56.0
pyee==13.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
rq==2.6.1
six==1.17.0
typing_extensions==4.15.0
Werkzeug==3.1.4
